requirements.txt
flask==3.1.0
flask-socketio==5.5.1
flask-session==0.8.0
psycopg2-binary==2.9.10
redis==5.2.1
eventlet==0.37.0
gunicorn==23.0.0
werkzeug==3.1.0
prometheus_client==0.21.1
flask-limiter==3.8.0
requests==2.32.3
